end_user fk type integer to bigint

+++ db/migrate/20230604153030_create_workout_comments.rb
@@ -1,7 +1,7 @@
 class CreateWorkoutComments < ActiveRecord::Migration[6.1]
   def change
     create_table :workout_comments do |t|
-      t.references :end_user, type: :integer, null: false, foreign_key: true
+      t.references :end_user, type: :bigint, null: false, foreign_key: true
       t.references :post_workout, type: :integer, null: false, foreign_key: true
       t.text :comment,    null: false
