+++ db/migrate/20230604153135_create_meal_likes.rb
@@ -1,7 +1,7 @@
 class CreateMealLikes < ActiveRecord::Migration[6.1]
   def change
     create_table :meal_likes do |t|
-      t.references :end_user, type: :integer, foreign_key: true, null: false
+      t.references :end_user, type: :bigint, foreign_key: true, null: false
       t.references :post_meal, type: :integer, foreign_key: true, null: false
 
       t.timestamps
