+++ db/migrate/20230604153148_create_meal_comments.rb
@@ -1,7 +1,7 @@
 class CreateMealComments < ActiveRecord::Migration[6.1]
   def change
     create_table :meal_comments do |t|
-      t.references :end_user, type: :integer, null: false, foreign_key: true
+      t.references :end_user, type: :bigint, null: false, foreign_key: true
       t.references :post_meal, type: :integer, null: false, foreign_key: true
       t.text :comment,    null: false
